requirements.txt
absl-py==2.1.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.6.0
asn1crypto==0.24.0
asttokens==2.4.1
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.7.0
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.3.3
cbor==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.7
cmake==3.28.3
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
distlib==0.3.0
distro==1.5.0
distro-info==0.23+ubuntu1.1
Django==4.1.7
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
executing==2.0.1
filelock==3.0.12
flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==7.1.0
importlib-resources==6.4.0
incremental==16.10.1
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==3.2.0
jupyter-client==8.6.1
jupyter-core==5.7.2
keras==2.13.1
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
more-itertools==4.2.0
mpi4py==3.0.3
nest-asyncio==1.6.0
netifaces==0.10.4
numpy==1.24.3
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
paramiko==2.7.1
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.2.0
plotly==5.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.5.1
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygments==2.17.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
pyproj==2.5.0
PyQRCode==1.2.1
pyRFC3339==1.1
pyrsistent==0.16.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.14.0
python-snappy==0.5.3
PyTrie==0.2
pytz==2024.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.2
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
retrying==1.3.4
rsa==4.9
Rtree==0.9.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
stack-data==0.6.3
stevedore==1.32.0
system-service==0.3
systemd-python==234
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
texttable==1.6.2
threadpoolctl==3.4.0
tornado==6.4
traitlets==5.14.2
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.5.0
tzdata==2024.1
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.13
websocket-client==0.57.0
werkzeug==3.0.1
wrapt==1.16.0
wsaccel==0.6.2
xkit==0.0.0
zipp==3.18.1
zope.interface==4.7.1
